backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
alembic>=1.14.0
python-dotenv>=1.0.0
# Telegram Bot
python-telegram-bot>=21.0
httpx>=0.27.0
# iCal Import
icalendar>=6.1.0
python-dateutil>=2.9.0
# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
# Background Tasks
apscheduler>=3.10.4
